# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API for get news from fact checking websites",
        "title": "Fake Hub",
        "contact": {
            "name": "Igor Vieira de Mello",
            "email": "ivmello@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/crawl/lupa": {
            "post": {
                "description": "search on Agencia Lupa query param sent on body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crawl/lupa"
                ],
                "summary": "Agencia Lupa crawler",
                "parameters": [
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LupaPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "api.LupaPayload": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string"
                }
            }
        }
    }
}